+++ Sources/CLI/XCSnippet/RemoveSnippetsCommand.swift
@@ -7,9 +7,9 @@ struct RemoveSnippetsCommand: ParsableCommand {
 
     @Option(
         name: .shortAndLong,
-        help: "The snippet namespace to delete."
+        help: "The snippet namespace to delete. All the snippets are deleted if not specified."
     )
-    var namespace: String = "FABERNOVEL"
+    var namespace: String?
 
     public static let configuration = CommandConfiguration(
         commandName: "remove",
@@ -19,6 +19,6 @@ struct RemoveSnippetsCommand: ParsableCommand {
     // MARK: - ParsableCommand
 
     func run() throws {
-        try XCSnippetCLI().removeSnippets(for: XCSnippetNamespace(namespace))
+        try XCSnippetCLI().removeSnippets(namespace: namespace)
     }
 }