+++ Changelog.md
@@ -3,6 +3,18 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.1.3]
+
+### Fixed
+
+- Fixed relative URL parameters #7
+- Fixed empty code snippets folder #11
+
+### Updated
+
+- Updated `xcresource snippet remove`: it can not delete custom snippets
+- Updated CLI messages
+
 ## [0.1.2]
 
 ### Updated